Change the name of the config class to be more consistent
