+++ bungee/src/main/java/us/myles/ViaVersion/bungee/platform/BungeeViaConfig.java
@@ -1,40 +1,58 @@
-package us.myles.ViaVersion.sponge.platform;
+package us.myles.ViaVersion.bungee.platform;
 
-import org.spongepowered.api.asset.Asset;
-import org.spongepowered.api.plugin.PluginContainer;
 import us.myles.ViaVersion.api.ViaVersionConfig;
+import us.myles.ViaVersion.api.protocol.ProtocolVersion;
+import us.myles.ViaVersion.bungee.providers.BungeeVersionProvider;
 import us.myles.ViaVersion.util.Config;
 
 import java.io.File;
 import java.net.URL;
-import java.util.Arrays;
-import java.util.List;
-import java.util.Map;
-import java.util.Optional;
+import java.util.*;
 
-public class SpongeConfigAPI extends Config implements ViaVersionConfig {
-    private static List<String> UNSUPPORTED = Arrays.asList("anti-xray-patch", "bungee-ping-interval", "bungee-ping-save", "bungee-servers", "velocity-ping-interval", "velocity-ping-save", "velocity-servers", "quick-move-action-fix");
-    private final PluginContainer pluginContainer;
+public class BungeeViaConfig extends Config implements ViaVersionConfig {
+    private static List<String> UNSUPPORTED = Arrays.asList("nms-player-ticking", "item-cache", "anti-xray-patch", "quick-move-action-fix", "velocity-ping-interval", "velocity-ping-save", "velocity-servers");
 
-    public SpongeConfigAPI(PluginContainer pluginContainer, File configFile) {
+    public BungeeViaConfig(File configFile) {
         super(new File(configFile, "config.yml"));
-        this.pluginContainer = pluginContainer;
         // Load config
         reloadConfig();
     }
 
     @Override
     public URL getDefaultConfigURL() {
-        Optional<Asset> config = pluginContainer.getAsset("config.yml");
-        if (!config.isPresent()) {
-            throw new IllegalArgumentException("Default config is missing from jar");
-        }
-        return config.get().getUrl();
+        return BungeeViaConfig.class.getClassLoader().getResource("assets/viaversion/config.yml");
     }
 
     @Override
     protected void handleConfig(Map<String, Object> config) {
-        // Nothing Currently
+        // Parse servers
+        Map<String, Object> servers;
+        if (!(config.get("bungee-servers") instanceof Map)) {
+            servers = new HashMap<>();
+        } else {
+            servers = (Map) config.get("bungee-servers");
+        }
+        // Convert any bad Protocol Ids
+        for (Map.Entry<String, Object> entry : new HashSet<>(servers.entrySet())) {
+            if (!(entry.getValue() instanceof Integer)) {
+                if (entry.getValue() instanceof String) {
+                    ProtocolVersion found = ProtocolVersion.getClosest((String) entry.getValue());
+                    if (found != null) {
+                        servers.put(entry.getKey(), found.getId());
+                    } else {
+                        servers.remove(entry.getKey()); // Remove!
+                    }
+                } else {
+                    servers.remove(entry.getKey()); // Remove!
+                }
+            }
+        }
+        // Ensure default exists
+        if (!servers.containsKey("default")) {
+            servers.put("default", BungeeVersionProvider.getLowestSupportedVersion());
+        }
+        // Put back
+        config.put("bungee-servers", servers);
     }
 
     @Override
@@ -153,12 +171,12 @@ public class SpongeConfigAPI extends Config implements ViaVersionConfig {
 
     @Override
     public boolean isItemCache() {
-        return getBoolean("item-cache", true);
+        return false;
     }
 
     @Override
     public boolean isNMSPlayerTicking() {
-        return getBoolean("nms-player-ticking", true);
+        return false;
     }
 
     @Override
@@ -209,6 +227,35 @@ public class SpongeConfigAPI extends Config implements ViaVersionConfig {
     @Override
     public boolean isMinimizeCooldown() {
         return getBoolean("minimize-cooldown", true);
+    }
+
+    /**
+     * What is the interval for checking servers via ping
+     * -1 for disabled
+     *
+     * @return Ping interval in seconds
+     */
+    public int getBungeePingInterval() {
+        return getInt("bungee-ping-interval", 60);
+    }
+
+    /**
+     * Should the bungee ping be saved to the config on change.
+     *
+     * @return True if it should save
+     */
+    public boolean isBungeePingSave() {
+        return getBoolean("bungee-ping-save", true);
+    }
+
+    /**
+     * Get the listed server protocols in the config.
+     * default will be listed as default.
+     *
+     * @return Map of String, Integer
+     */
+    public Map<String, Integer> getBungeeServerProtocols() {
+        return get("bungee-servers", Map.class, new HashMap<>());
     }
 
     @Override
