+++ bungee/src/main/java/us/myles/ViaVersion/bungee/service/ProtocolDetectorService.java
@@ -6,7 +6,7 @@ import net.md_5.bungee.api.ServerPing;
 import net.md_5.bungee.api.config.ServerInfo;
 import us.myles.ViaVersion.BungeePlugin;
 import us.myles.ViaVersion.api.Via;
-import us.myles.ViaVersion.bungee.platform.BungeeConfigAPI;
+import us.myles.ViaVersion.bungee.platform.BungeeViaConfig;
 import us.myles.ViaVersion.bungee.providers.BungeeVersionProvider;
 
 import java.util.HashMap;
@@ -26,7 +26,7 @@ public class ProtocolDetectorService implements Runnable {
 
     public static Integer getProtocolId(String serverName) {
         // Step 1. Check Config
-        Map<String, Integer> servers = ((BungeeConfigAPI) Via.getConfig()).getBungeeServerProtocols();
+        Map<String, Integer> servers = ((BungeeViaConfig) Via.getConfig()).getBungeeServerProtocols();
         Integer protocol = servers.get(serverName);
         if (protocol != null) {
             return protocol;
@@ -61,8 +61,8 @@ public class ProtocolDetectorService implements Runnable {
                     // Ensure protocol is positive, some services will return -1
                     if (serverPing.getVersion().getProtocol() > 0) {
                         detectedProtocolIds.put(key, serverPing.getVersion().getProtocol());
-                        if (((BungeeConfigAPI) Via.getConfig()).isBungeePingSave()) {
-                            Map<String, Integer> servers = ((BungeeConfigAPI) Via.getConfig()).getBungeeServerProtocols();
+                        if (((BungeeViaConfig) Via.getConfig()).isBungeePingSave()) {
+                            Map<String, Integer> servers = ((BungeeViaConfig) Via.getConfig()).getBungeeServerProtocols();
                             Integer protocol = servers.get(key);
                             if (protocol != null && protocol == serverPing.getVersion().getProtocol()) {
                                 return;
