+++ sponge/src/main/java/us/myles/ViaVersion/SpongePlugin.java
@@ -17,7 +17,6 @@ import org.spongepowered.api.plugin.PluginContainer;
 import org.spongepowered.api.scheduler.Task;
 import org.spongepowered.api.text.serializer.TextSerializers;
 import us.myles.ViaVersion.api.Via;
-import us.myles.ViaVersion.api.ViaAPI;
 import us.myles.ViaVersion.api.command.ViaCommandSender;
 import us.myles.ViaVersion.api.configuration.ConfigurationProvider;
 import us.myles.ViaVersion.api.platform.TaskId;
@@ -32,10 +31,8 @@ import us.myles.ViaVersion.util.GsonUtil;
 
 import java.io.File;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.List;
 import java.util.UUID;
-import java.util.concurrent.TimeUnit;
 import java.util.logging.Logger;
 
 @Plugin(id = "viaversion",
@@ -58,7 +55,7 @@ public class SpongePlugin implements ViaPlatform {
     @Getter
     private SpongeViaAPI api = new SpongeViaAPI();
     @Getter
-    private SpongeConfigAPI conf;
+    private SpongeViaConfig conf;
 
     @Getter
     private Logger logger;
@@ -68,7 +65,7 @@ public class SpongePlugin implements ViaPlatform {
         // Setup Logger
         logger = new LoggerWrapper(container.getLogger());
         // Setup Plugin
-        conf = new SpongeConfigAPI(container, defaultConfig.getParentFile());
+        conf = new SpongeViaConfig(container, defaultConfig.getParentFile());
         SpongeCommandHandler commandHandler = new SpongeCommandHandler();
         game.getCommandManager().register(this, commandHandler, "viaversion", "viaver", "vvsponge");
         getLogger().info("ViaVersion " + getPluginVersion() + " is now loaded!");
